Growroot working
Adapted growroot script from debian cloud-initramfs-growroot to run
before mounting of the root partion. The Debian script does not check
whether other partions (like a read only usr) partition are mounted
and fails. This cannot happen here.

As a downside the script cannot (easily) report what it has done so it
will attempt to grow the partion the next time it boots (where it reports
that it will do nothing).