generate and add variant-specific manifest
The contract between CICD build pipeline and `bin/garden-build.sh` is
defined in such a way that CICD build pipeline will accept any file
output by `bin/garden-build-sh`.

As the set of files is thus not part of the contract / subject to
change, generate an accompanying manifest (at a hard-coded, defined
position (prefix + hardcoded suffix)), so that all files belonging to
the build result are identifyable easily.

To be extended, e.g. to add metadata to special build result files.