Make ssh configuration more secure and autologout
For ssh explicitly disable problematic feature and only use secure
ciphers.

Autologout is more problematic. The settings in sshd_config did not
work with ClientAliveInterval set to 60 (in contract to what is documented
all of the web). Adding the desired value of 600 seconds:

ClientAliveInterval 600
ClientAliveCountMax 0

Therefore also setting the TMOUT variable to 10 minutes as well. This
however does only work with bash, Zsh and ksh. It does not work bith
bourne shell.