vmware: Support cloud-init datasource GuestInfo
This patch adds support for the Cloud-Init Datasource,
VMwareGuestInfo, to the VMware image. The files in this
commit were copied from release v1.4.1 of the datasource
from https://github.com/vmware/cloud-init-vmware-guestinfo.

The datasource has a dependency on the netifaces library,
and so the package python3-netifaces was added to the package
file for the vmware image.