ci: build step-image on the fly
In order to reduce the size of our job-image, we'd like to
remove all non-gardener dependencies from our image-build.
In order for the garden-linux build (and publishing) to still
function, the Pipeline now builds an image that is equivalent to
the one used before and then uses it.
Since we still need an image to run the initial clone and
image-builds this won't get rid of all uses of the old job-image,
but what remains can be serviced by the new reduced job-image.