enable ssh-rsa as HostKeyAlgorithm
ssh-rsa was dropped from OpenSSH 8.8 as it is considered insecure
(ref https://www.openssh.com/txt/release-8.8). While it is reasonable
for SSH clients not to use this algorithm anymore, the SSH server sshd
should still advertise it as valid algorithm for host-key verification
since there are still clients out there (such as Azure's Certification
Test Tool for Azure Certified) which will simply be unable to establish
a connection without.