hack: enable ssh-rsa for Az Marketplace
Azure Marketplace certification uses an automated service which not only
requires ssh-rsa for HostKeyAlgorithms but also in
PubkeyAcceptedAlgorithms.
Since default support for ssh-rsa was dropped from OpenSSH 8.8 due to
security concerns with the embedded SHA-1 signature algorithm, we have
to explictly enable it here just for the Azure feature so that the image
can pass Azure Marketplace certification.
Once Azure updates its service to be compatible with ssh-rsa-256 or
ssh-rsa-512, this commit should be reverted.