fix(ci): github actions does not set pipefail
The bash set builtin "pipefail" is not set in run-actions of github actions.
See GitHub Docs[1] and also here [3].

An addtional keyword "shell" can be specified to tell the run-actions
explicitly what shell to use. Using "shell: bash" sets the defaults to

  bash --noprofile --norc -eo pipefail {0}

if "shell: bash" is not set, pipefail is also not set.

Which leads to the following problem, fixed with this commit:

The code:
    false | tee some/log/destination

will return a non-error code 0, if tee was successfull.
This is obviously not what we want when executing a test.

References:
[1] https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#exit-codes-and-error-action-preference
[2] https://www.gnu.org/software/bash/manual/html_node/The-Set-Builtin.html
[3] https://github.com/actions/runner/blob/main/docs/adrs/0277-run-action-shell-options.md