fix(azure test): make nopasswd explicit
* A test checks the permissions returned
via "sudo -l -U $user" for all available users.
To perform this tests, the admin user must be
able to execute it without entering a password
in a test setup.
For all other cloud providers we do this implicitly
already via cloud-init or something else.
For Azure this is not managed by garden linux,
but by the user (in this example, the platform tests.)