test(azure): increase threshold times for azure
* azure requires sometimes more time during boot
* The included timing tests have the purpose to
alert us if anything is out of the ordinary.
For azure, long times for these tests are not
unusual.
* Presumably there are variables in time during
registration of azure VM that depend on the
azure cloud registration backends. However,
this needs to be investigated and is not proven.
Until then, we decide to increase the acceptance
threshold for azure, since the focus of the
platform tests are not performance tests.
Performance tests would require a more thorough
setup with multiple runs to get meaningful
statistical data.
We decided against nightly performance tests to
save computing resources.