replace subprocess calls with boto3 api calls
The calls to the external aws cli program were replaced with calls to
the AWS boto3 API - the general functionality of this tools remains
unchanged.

An S3 bucket, which is required for the image upload can be created
on-the-fly if not bucket name is given and will be automatically removed
again once the image snapshot was created.