Exclude firewall feature from gardener feature
The firewall feature would configure nftables with a default
policy to DROP everything on the input chain that is not SSH.
This severely harms any kind of Kubernetes operation as it
interferes with CNI. Therefore, the firewall feature must not
be used with the Gardener feature.

Additionally: a test that can check for the presence of
undesired files (like /etc/nft.d and /etc/nftables.conf in this
case) and which will fail if these files were found in the
image.