fix(ci): nightly for 934 must be triggered with version
* if version is not explicitly set, the default "now" is used.
The "now" parameter uses the current date and ignores the VERSION file.
Thus, the 934 build has the wrong version input.

* This commit explicitly sets the version as input, by reading the
content of the VERSION file used in the rel-934 branch.

* For a nightly and development the default "now" does make sense.
However, for building legacy, we need to tell the builder what version
to use. The VERSION file was always the source for that, therefore
we may want to control the default via the VERSION file.