Add a script to update the dev env on demand (#1809)
This PR addresses an issue with the dev environment:

Previously, the dev environment automatically always used the latest debian testing image available which changes every day. The issue is that debian testing might be broken on some days, so we never know if this will be a working environment. The current example for this issue is [crun and the linux kernel](https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1053821) where a new kernel broke crun, and crun was fixed upstream but not in debian testing for a while.

The alternative approach is to use specific debian testing images as introduced in #1796. This has the upside of producing a more reliable environment, but the downside of additional maintenance effort as this image will have to be updated from time to time to avoid it getting too old.

This PR introduces a script to automate the update process, so we can have reliable environments but still update them regularly (probably once or twice a month) without much effort. Running the script will get the most recent debian testing image and rewrite the lima manifest yaml file. Note that testing if the new image works for building Garden Linux is not automated and has to be done before merging changes produced by this script.