test: handle pam_nologin during connection attempt
In some rare cases, the Garden Linux platform test framework
attempts to connect to a machine during its boot-up in that small
time frame when SSH already accepts connections but PAM refuses
them with the "System is booting up" message. In this case, the
tests should not be aborted with an error but the connection should
be retried.