fix(test_proc): rewrite test to check if proc/ is empty in image artefact
This commit changes the test_proc to check if there are any files in the image artefact
starting with "proc/". The previous implementation of test_proc.py checked for any files in the image artefact
containing the string "proc/". A file in the artefact containing "remoteproc/" in the file path causes the test to fail, creating a false alert.