cloud-init network config for OpenStack baremet to circumvent the race.l
To configure network bonds/trunks/LACP on OpenStack baremetal
machines, cloud-init must be allowed to configure the network
and hand over the network config to netplan so that it can
take care of initialising the bond* interface before any
dhcp request is made.

Also, since some network interfaces take a little longer to
initialise, this can introduce a race between them and
cloud-init-local which will fail if the network interfaces listed
in a network_config.json are not present yet. The bnxt_en network
driver is placed into initrd so that it loads and initialises at
a very early stage to circumvent the race.