Resize metal /efi to 300MiB
Garden Linux images have a default /efi size of 64MiB. This is not sufficient for more than one kernel, which often prevents Kernel Updates from working. We experienced this with our images and customers also experienced it. I propose to enlarge the /efi partition to have no issues if the need for two (or even three) images arises. 

Although this is against the "create a small/minimal immutable image" rule, but i think practically this would help many users in the future.