Remove assertion for capabilities on 'ping' (#2530)
`iputils` was updated in debian testing with a newer version of ping that does not require setting `cap_net_raw`.
As the requirement of the package has changed, the test now does not test for the right thing. Now we want to test that the capability is not set.

For reference: [This is the related Debian bug](https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1008281)

[this is the change log entry](https://metadata.ftp-master.debian.org/changelogs/main/i/iputils/testing_changelog)

```
* ping: Add Recommends on linux-sysctl-defaults and stop setting additional
     capabilities (Closes: #[1008281](https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1008281))
```

Also fixes a few typos that came up while working on it.