Remove mark manual for packages not in the repo (#2580)
The removed code marks some packages as 'manually installed' as described in #1305.
This now breaks the build as some of those packages are not in the repo anymore.
This code was a workaround for an issue with debootstrap in 2022, I don't think we need this anymore, so we can remove it to make the build work again.

Fixes #2579