Automate package upgrade release notes using glvd (#2662)
Get list of fixed CVEs, grouped by upgraded package.
This automates some of the work needed when a new release is done.

Note that the output is known to be not perfect, it depends on the data available in glvd. Manual editing of the output is anticipated, but it should still reduce the manual effort needed on creating a new release.

Fixes https://github.com/gardenlinux/gardenlinux/issues/2653