fix: test chroot script to ignore unrelated symlink in container buildcontext (#2759)
* fix: make ./test work again alongside openTofu setup

The script ./test calls podman build with folder ./tests as build context.
The ./test will fail without this patch, because an unrelated symlink
points outside the buildcontext which the podman build checks and throws an
Error that the symlink is broken.

error: invalid symlink "/var/tmp/libpod.../build/platformSetup/tofu/cert" -> "../../../cert"

This commit tells podman to ignore that file