Clean up the conftest.py into smaller pieces (#2706)
In this commit, we split the configuration into smaller sized configuration files to allow better maintaining the code.

During my testing on the last day, I noticed some misconfigurations happening within the conftest.py file. 
For instance, functions to detect for firecracker or local were added more than once and with out-dated 
configuration settings. Also, adding more entries was error prune, since details were earlier lost.

The split is implemented by treating all the configuration settings as pytest_plugins variable. 
Please note: This will only work since we're using a top-level conftest.py, 
we can't use this in any lower level of it.

This commit squashes several commits together:

* Cleanup conftest
* Move conftest into a subfolder.
* Allow implicy import of conftest folder.
* Move all platforms into a seperate file
* Remove uncessary import
* Correct import from logger to logging
* fix typo
* Move miscellaneous settings into a file of it's own.
* Add all platforms from conftest to conftests/platforms
* Disable uci test for the moment.
* Move remaining configure for features and elements
* Remove dupliated entry for local, re-enable test_oci.
* Update tests/conftests/features.py
* Update tests/conftests/features.py
* Update tests/conftests/platforms.py
* Fix grammer
* Fix grammer
* Include feedback
* Move Iaas related setting into iaas.py conftest, rename kvm relate iaas to qemu.
* Rename iaas to provisioner, rename kvm to qemu for local setup.
* Add --provisioner flag, set warning for --iaas flag.
* Replace most iaas with provisioner.
* Add function to deal with local tests with iaas.
* Some addional change from iaas to provisioner
* Update the configuration case for iaas
* Update if to not considere unnecessary iaas paramenter
* Revert "Update if to not considere unnecessary iaas paramenter"
* Revert "Update the configuration case for iaas"
* Revert "Some addional change from iaas to provisioner"
* Revert "Add function to deal with local tests with iaas."
* Revert "Replace most iaas with provisioner."
* Revert "Add --provisioner flag, set warning for --iaas flag."
* Some PEP8 love and improved documentation.

The scope of the task became bigger, and we noticed problems with the 
usage of certain terms like platform, container and iaas. This will be addressed
in a different issue.


Relates: #2784

---------

Co-authored-by: Vincent Riesop <v.riesop@sap.com>