feature/test su (#2823)
In this PR, we added a test to ensure that access to `su` is restricted by the OS. 

The test will read the file from `/etc/pam.d/su` and check if it has the configuration parameter we expect.
A user needs to be configured to be admitted to use `su`. By default, this is realized by a specify group, the `wheel` group. We have to ensure that this behavior is enabled. 

This way, the restricting of what target identities are allowed to elevate is done via the group `wheel` and by not fizzing around with the PAM configuration files. 

This commit squashes the following commits: 

* Test that the access to su is restricted.
* Add sudo permission
* Update features/base/test/test_su.py
* Fix sudo, make it more readiable
* Update doc string
* use black

Closes: #2787 


Co-authored-by: Moritz Kirchner <moritz.kirchner@sap.com>