fix test: disable test_clocksource on qemu without kvm support (#2941)
The test

	base/test/test_time_config.py::test_clocksource

will fail with

	AssertionError: unknown hypervisor qemu

This happens because systemd-detect-virt will output "qemu" in case we start tests via
qemu target, e.g.:

	make --directory=tests aws-gardener_prod-amd64-qemu-test-platform

In order to exclude this test for this test target, we need to add new fixtures for "qemu" iaas.
As defined in tests/conftest.py::testconfig, iaas can be set to qemu.

This commit adds:
- fixtures to exclude/include qemu iaas
- excludes test_time_config.py::test_clocksource for qemu platform