Generate lima manifest automatically when building (#3118)
This is a convenience feature to make it more simple/easy to spin up lima vms based on custom built images. Having the yaml manifest with the correct file name available is helpful to create vm easily using the `limactl` tool.

This PR is related to https://github.com/gardenlinux/gardenlinux/issues/2965