Enhance test markers to include optional 'reasons' argument (#3320)
This PR allows us to better keep track of the reason for certain markers. This field is optional, so for obvious cases it can be left empty, but I think in most cases it is actually useful to document the reasoning for setting a certain marker.

Fixes https://github.com/gardenlinux/gardenlinux/issues/3319