Add FIPS support for Garden Linux (#3515)
In this commit, we added a basis for FIPS. It will configure the Linux kernel to be booted with the fips=1 parameter; it has all the files necessary by gnutls, libgcrypt and openssl to be in the FIPS mode.

As of now, it comes with limitations. gnutls is broken when performing the self-test. And we can't use the HTTPS with the apt mirror as of now, since the connection breaks when downloading packages. Both aspects will be addressed in separate tickets (#3626, #3581 )

Related: #3436