Document decision to ban rebase merge to ensure traceability of commits (#3672)
Following the discussion we had in the daily today, I've drafted a decision for disallowing rebase merge for the benefit of traceability.

This PR will stay open at least until our next team meeting where I intend to have a decision on this. If you have feedback, feel free to comment here in the PR, or write to me directly.