Port 'security_id' mark to new test framework (#3708)
We had used `security_id` in tests without having the mark implemented in the new test framework. This PR ports the old implementation of the mark into the new framework, in the form of a plugin. This allows us to remove the warnings without commenting out the marks, and should be the way forward.

Port of #2477
Related to #3691
Fixes #3707