Disable DSA, since it shound't be longer used. (#3804)
In this commit, we disable the generation of DSA for the chroot environment. 

During enablement of the [FIPS flavor](https://github.com/gardenlinux/gardenlinux/pull/3744), the old testing framework was running. The tests failed to boot up the chroot environment for a FIPS-enabled environment. The error is SUPER easy. We still generate a DSA key, which fails. I removed the necessary line. DSA shouldn't be used anyway since OpenSSH has removed it as well. It fails in FIPS, since FIPS disables the DSA function, which OpenSSH depends on to compute.

Resolves: #3803
Relates: #3436