Set OPENSSL_MODULES to load fips.so (#3797)
In this commit, we set OPENSSL_MODULES to load fips.so for the test-ng.

While developing a test for GnuTLS to ensure that we have the correct HMAC computed, I noticed a 'bug' in OpenSSL/Python/Testing Framework. It's an oddity. The problem is that the Python version can't load the `fips.so`, by setting this as an environment variable, it resolves the issue.  

Resolves: #3797
Relates: #3701, #3436