Feature/3701 tests for fips (#3752)
In this commit, we add some basic tests for GnuTLS, libgcrypt, and the kernel regarding the FIPS mode. It validates only that the necessary prerequisites are present to enable the FIPS mode. 

When the tests succeed, Garden Linux can be used with FIPS mode. It does have a simple test to ensure the kernel is in FIPS mode. It does not test any of the other libraries (yet). Besides the kernel test, all tests can be conducted within a chroot environment.
 

These commits were squashed: 

* Initial test for fips
* Add basic tests for gnutls fips configuration
* Add basic tests for libgcrypt fips configuration
* Add basic tests for kernel fips configuration
* Add runtime tests for kernel fips configuration
* Some black love
* Some grammer corrections
* Some isort love
* Add marks to ensure it will be tested only with _fips
* DSA is not longer support, remove generation of keys

Relates: #3701, #3436