Check adr numbering (#3929)
Introduce a new workflow to ensure ordering of ADRs is correct (no id is reused, no gaps between ids).

Discovered an issue with 'ADR 0013: System State Diffing in tests-ng', fixing this by putting it to the back to avoid having to rename all ADRs that were created since then.

From now on this should not happen with this workflow.