Add more tests for GnuTLS (#3899)
In this commit, we add some more advanced tests for the FIPS feature.

- test_gnutls_fips_dot_hmac_file_is_presented
- test_gnutls_fips_dot_hmac_file_is_vaild
- test_kernel_has_fips_entry_in_procfs

GnuTLS will perform a self-check based on the FIPS requirements. A file that contains an HMAC needs to be present on the target system. The tests ensures that the file was installed and contains a valid HMAC. One problem in shipping GnuTLS packages was that the computed HMAC was not valid for the shipped version of the library since the HMAC was computed for an unstripped version. 

Lastly, we clean up some of the other tests to be more readable and use a plugin instead of reading the kernel cmd ourselves.

The following commits were squashed in:  

* Add more tests for GnuTLS
* Remove misplace ignore for pyright.
* Update tests-ng/test_fips.py
* Remove trace of refactor code
* Simplify tests for file check and clean up
* Using os.path.ispath for test_gnutls_fips_dot_hmac_file_is_presented
* Remove unused import
* Refactor the test_kernel_was_boot_with_fips_mode, add extra test for procfs
* Some love for black.
* Sorting imports
* Add missing import for pyright
* Fix the path for gnutls test
* The cleanup was to much, fix it 😛
* Feedback
* Remove newline from test

---------

Co-authored-by: Moritz Kirchner <moritz.kirchner@sap.com>

Relates: #3898