GnuTLS FIPS test needs a booted VM. (#4067)
In this commit, we resolve a problem that the GnuTLS test for FIPS needs a booted system.

GnuTLS needs the kernel to be booted in FIPS mode to work. You can force GnuTLS to be in FIPS mode `GNUTLS_FORCE_FIPS_MODE`, but we do not do this with this test. 

> There are two distinct library states with regard to FIPS140-2: the FIPS140-2 mode is installed if /etc/system-fips is present, and the FIPS140-2 mode is enabled if /proc/sys/crypto/fips_enabled contains ’1’, which is typically set with the “fips=1” kernel command line option.

https://www.gnutls.org/manual/html_node/FIPS140_002d2-mode.html

* GnuTLS FIPS test needs a booted VM.
* Add named parameter to the pytest.mark.booted
* Add missing fips mark to test_libgcrypt_is_in_fips_mode